refactor : PositiveNumbers 일급컬렉션 적용  - StringAddCalculator 비즈니스 로직 PositiveNumbers 클래스로 이동  - PositiveNumbers 클래스 상수적용
